Remove spurious "GROUP BY" clause
I'm guessing this was introduced because we have to use a HAVING
clause (as opposed to a WHERE clause) to filter on the computer
last_event column, and whoever added it didn't realize you can still
use HAVING without GROUP BY. It's also unclear to me whether using
GROUP BY without any aggregation functions is valid.

In any case, the GROUP BY clause was breaking MariaDB compatibility,
so removing it as unnecessary.