Reuse OIDC token verifier
The OIDC token verifier is safe for concurrent use, and reusing it
allows us to cache the JWKS credentials and avoid extra network calls
for each login.

Logins are infrequent enough today that this isn't a big deal, but it
may allow us to use the OIDC id_tokens as bearer tokens directly.